+++ app/api/hello/route.js
@@ -1,0 +1,3 @@
+export async function GET(_request) {
+  return Response.json({ name: 'John Doe' })
+}
