+++ eslint.config.js
@@ -1,0 +1,11 @@
+import nextConfig from "eslint-config-next";
+
+// eslint-disable-next-line import/no-anonymous-default-export
+export default [
+  ...nextConfig,
+  {
+    rules: {
+      // Add any custom rules here
+    },
+  },
+];
