+++ next.config.js
@@ -1,3 +1,6 @@
-module.exports = {
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   reactStrictMode: true,
 }
+
+export default nextConfig
